Don't cast sizeof() results to unsigned when passing to Xmalloc/Xcalloc
sizeof() returns size_t, malloc() & calloc() expect sizes in size_t,
don't strip down to unsigned int and re-expand unnecessarily.

+++ src/xcms/cmsCmap.c
@@ -250,8 +250,7 @@ _XcmsAddCmapRec(
 {
     XcmsCmapRec *pNew;
 
-    if ((pNew = Xcalloc(1, (unsigned) sizeof(XcmsCmapRec)))
-	    == NULL) {
+    if ((pNew = Xcalloc(1, sizeof(XcmsCmapRec))) == NULL) {
 	return((XcmsCmapRec *)NULL);
     }
 
@@ -301,7 +300,7 @@ _XcmsCopyCmapRecAndFree(
 	pRec_copy =_XcmsAddCmapRec(dpy, copy_cmap, pRec_src->windowID,
 		pRec_src->visual);
 	if (pRec_copy != NULL && pRec_src->ccc) {
-	    pRec_copy->ccc = Xcalloc(1, (unsigned) sizeof(XcmsCCCRec));
+	    pRec_copy->ccc = Xcalloc(1, sizeof(XcmsCCCRec));
 	    memcpy((char *)pRec_copy->ccc, (char *)pRec_src->ccc,
 		   sizeof(XcmsCCCRec));
 	}
